apps/ai/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.26.2
torch==2.1.0
# Grounding DINO uses model_type "mm-grounding-dino" — only in transformers>=4.46
transformers>=4.46.0,<5.0.0
httpx>=0.24.0,<0.25.0
# Optional: OCR on detected "brand text" / "other text" regions
# pip install pytesseract  # and system tesseract-ocr

